introduce workspace struct
