+++ src/projects.rs
@@ -1,4 +1,6 @@
-use anyhow::Result;
+use crate::{node, nx};
+use anyhow::{bail, Result};
+use log::debug;
 use std::path::Path;
 
 /// A trait for defining a project.
@@ -9,10 +11,20 @@ pub trait Project {
         Self: Sized;
 }
 
-pub fn is_project_dir(path: &Path) -> bool {
-    path.is_dir()
-        && (
-            path.join("project.json").is_file() || path.join("package.json").is_file()
-            // || path.join("Cargo.toml").is_file()
-        )
+pub fn get_project(workspace_root: &Path, project_path: &str) -> Result<Box<dyn Project>> {
+    let project_root = workspace_root.join(project_path);
+    let project_json_path = project_root.join("project.json");
+    let package_json_path = project_root.join("package.json");
+
+    if project_json_path.is_file() {
+        let nx_proj = nx::NxProject::load(workspace_root, project_path)?;
+        debug!("{:?}", nx_proj);
+        Ok(Box::new(nx_proj))
+    } else if package_json_path.is_file() {
+        let node_proj = node::NodeProject::load(workspace_root, project_path)?;
+        debug!("{:?}", node_proj);
+        Ok(Box::new(node_proj))
+    } else {
+        bail!("Could not find 'project.json' or 'package.json' in the project directory");
+    }
 }
